Split unit tests into per-class files with expanded coverage
Replace the monolithic test/unit/api_surface_tests.cpp with focused
per-class test files, each building into its own standalone binary:

- api_spacegroup.cpp: SpaceGroup (settings, reflection conditions,
  symmetry operations, asymmetric unit)
- api_crystal.cpp: ScatteringPowerAtom, Crystal, UnitCell, Scatterer,
  Atom
- api_molecule.cpp: Molecule (atoms/bonds, angles/dihedrals, formula,
  log-likelihood)
- api_scattering.cpp: ScatteringData, DiffractionDataSingleCrystal,
  including RAD_XRAY/RAD_NEUTRON/RAD_ELECTRON coverage
- api_powderpattern.cpp: PowderPattern, PowderPatternBackground,
  PowderPatternDiffraction (incl. MuR absorption), ScatteringCorr
  subclasses (Lorentz/Polarization/SlitAperture), and all
  ReflectionProfile* classes
- api_cif.cpp: CIF import, data fields, coordinate conversion
- api_optimization.cpp: RefinablePar/RefinableObj, OptimizationObj/
  MonteCarloObj, LSQNumObj
- api_indexing.cpp: PeakList, CellExplorer
- ground_truth.cpp: standalone dump-ground-truth utility, kept
  separate from the test binaries per request

Shared helpers (Check/CheckNearAbs, tracing, ground-truth loaders/
comparators, crystal/profile builders) are factored into
test_common.h, included by each test file; since every .cpp still
compiles into its own binary this avoids ODR issues without
duplicating ~250 lines of logic 8 times.

Along the way:
- Fixed incorrect test assumptions (GetSymmetric(0,..) is not
  necessarily the identity position; AsymmetricUnit::IsInAsymmetricUnit
  has a pre-existing inverted bound check, so the new test only
  checks bounds validity/forwarding rather than semantic inclusion).
- Avoided LSQNumObj::Refine()'s fragile eigenvalue path for a minimal
  2-parameter synthetic background fit; test residual/obs/calc/weight
  vectors directly instead.
- Reworked the ReflectionProfileDoubleExponentialPseudoVoigt test to
  attach a real UnitCell (this profile divides by d_hkl, so a cell-less
  instance produces non-finite values) and avoided
  GetFullProfileWidth()'s search loop, whose d-spacing scale convention
  doesn't match GetProfile()'s and can walk out of bounds for arbitrary
  instrument parameters.

Updated test/unit/gnu.mak, run_unit_tests.sh and test/.gitignore for
the new 11-binary build. Full suite (45 cases) passes locally via
./run_unit_tests.sh.
